Split off tournament clock to own module
